Read metadata from file instead of variable
Previously it'd echo the metadata JSON into jq two times per chapter which
spammed the output of bash -x with useless information and probably isn't great
performance-wise.

I did this because this script used to not store the metadata on-disk but since
it does now, we can just read it from the file.